+++ .storybook/main.js
@@ -1,5 +1,7 @@
+const path = require('path');
+
 module.exports = {
-    stories: ['../src/**/*.stories.(js|mdx)'],
+    stories: ['../src/**/*.(stories|visual).(js|mdx)'],
 
     addons: [
         '@storybook/addon-knobs/register',
@@ -8,8 +10,7 @@ module.exports = {
         '@storybook/addon-storysource/register',
         '@storybook/addon-viewport/register',
         '@storybook/addon-docs',
-        'storybook-addon-i18n/register',
-        './.storybook/custom/register.js'
+        'storybook-addon-i18n/register'
     ],
 
     webpackFinal: async(config) => {
@@ -18,6 +19,12 @@ module.exports = {
             loaders: [require.resolve('@storybook/source-loader')],
             enforce: 'pre'
         });
+
+        config.module.rules.push({
+            test: /\.scss$/,
+            use: ['style-loader', 'css-loader', 'sass-loader'],
+            include: path.resolve(__dirname, '../'),
+          });
 
         return config;
     }