+++ src/InputGroup/InputGroup.js
@@ -5,7 +5,7 @@ import InputGroupAddon from './_InputGroupAddon';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 
-/** The **Input Group** includes form inputs with add-ons that allow the user to
+/** An **InputGroup** includes form inputs with add-ons that allow the user to
 better understand the information being entered. */
 class InputGroup extends Component {
     constructor(props) {
@@ -40,7 +40,8 @@ class InputGroup extends Component {
 
         const getClassName = (child) => classnames(
             {
-                'fd-input-group__input': child.type.displayName !== InputGroupAddon.displayName
+                'fd-input-group__input': child.type.displayName !== InputGroupAddon.displayName &&
+                    !child.props.className?.includes('fd-tokenizer')
             },
             child.props.className
         );