+++ src/InputGroup/__stories__/InputGroup.stories.js
@@ -53,7 +53,7 @@ export const disabled = () => (
 );
 
 export const validationState = () => (
-    <div className='fr-container'>
+    <div className='fddocs-container'>
         <InputGroup validationState={{ state: 'error', text: 'Test validation state' }}>
             <InputGroup.Addon>$</InputGroup.Addon>
             <FormInput />
@@ -74,7 +74,7 @@ export const validationState = () => (
 );
 
 export const addonButton = () => (
-    <div className='fr-container'>
+    <div className='fddocs-container'>
         <InputGroup>
             <FormInput />
             <InputGroup.Addon isButton>
@@ -105,7 +105,7 @@ addonButton.story = {
 };
 
 export const addonText = () => (
-    <div className='fr-container'>
+    <div className='fddocs-container'>
         <InputGroup>
             <InputGroup.Addon>€</InputGroup.Addon>
             <FormInput placeholder='Type text here' />
@@ -130,7 +130,7 @@ addonText.story = {
 };
 
 export const addonIcon = () => (
-    <div className='fr-container'>
+    <div className='fddocs-container'>
         <InputGroup>
             <InputGroup.Addon>
                 <Icon glyph='globe' />
@@ -158,7 +158,7 @@ addonIcon.story = {
 
 export const textArea = () => (
     <InputGroup>
-        <FormTextarea>Default</FormTextarea>
+        <FormTextarea defaultValue='Default' />
         <InputGroup.Addon>
             <Icon glyph='hide' />
         </InputGroup.Addon>