+++ src/Time/Time.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import TimeItem from './_TimeItem';
 import React, { Component } from 'react';
 
-/** The **Time** component is used for a single time value. Multiple components can be used in the **Time Picker**
+/** A **Time** component is used for a single time value. Multiple components can be used in the **Time Picker**
 to assemble a clock time. A max of four will account for hours, minutes, seconds and meridiem of the day.
 It is rarely used on its own as a standalone component. */
 
@@ -85,10 +85,8 @@ class Time extends Component {
             format12Hours,
             id,
             disabled,
-            onUpdateTime,
             onChange,
             time: timeProp,
-            name,
             hoursUpButtonProps,
             hoursDownButtonProps,
             hoursInputProps,
@@ -218,28 +216,13 @@ class Time extends Component {
 
 Time.displayName = 'Time';
 
-Time.basePropTypes = {
-    /** Internal use only */
-    /** Set to **true** to use the 12-hour clock (hours ranging from 01 to 12) and to display a meridiem control */
-    format12Hours: PropTypes.bool,
-    /** Enables the input for hours */
-    showHour: PropTypes.bool,
-    /** Enables the input for minutes */
-    showMinute: PropTypes.bool,
-    /** Enables the input for seconds */
-    showSecond: PropTypes.bool,
-    /** Set to **true** to show up/down buttons for each input */
-    spinners: PropTypes.bool,
-    /** The time component values. Contains four properties: **hour** (with values from 01 to 12 when `format12Hours`
-     * is true or 00 to 23 when `format12Hours` is false), **minute** (with values from 00 to 59), **second** (with values from 00 to 59),
-     * **meridiem** (with values 0 for AM or 1 for PM) */
-    time: PropTypes.object
-};
-
 Time.propTypes = {
-    ...Time.basePropTypes,
     /** Set to **true** to mark component as disabled and make it non-interactive */
     disabled: PropTypes.bool,
+    /** Internal use only */
+    disableStyles: PropTypes.bool,
+    /** Set to **true** to use the 12-hour clock (hours ranging from 01 to 12) and to display a meridiem control */
+    format12Hours: PropTypes.bool,
     /** Additional props to be spread to the hours down `<button>` element */
     hoursDownButtonProps: PropTypes.object,
     /** Additional props to be spread to the hours `<input>` element */
@@ -267,12 +250,24 @@ Time.propTypes = {
     minutesInputProps: PropTypes.object,
     /** Additional props to be spread to the minutes up `<button>` element */
     minutesUpButtonProps: PropTypes.object,
+    name: PropTypes.string,
     /** Additional props to be spread to the seconds down `<button>` element */
     secondsDownButtonProps: PropTypes.object,
     /** Additional props to be spread to the seconds `<input>` element */
     secondsInputProps: PropTypes.object,
     /** Additional props to be spread to the seconds up `<button>` element */
     secondsUpButtonProps: PropTypes.object,
+    /** Enables the input for hours */
+    showHour: PropTypes.bool,
+    /** Enables the input for minutes */
+    showMinute: PropTypes.bool,
+    /** Enables the input for seconds */
+    showSecond: PropTypes.bool,
+    /** Set to **true** to show up/down buttons for each input */
+    /** The time component values. Contains four properties: **hour** (with values from 01 to 12 when `format12Hours`
+      * is true or 00 to 23 when `format12Hours` is false), **minute** (with values from 00 to 59), **second** (with values from 00 to 59),
+      * **meridiem** (with values 0 for AM or 1 for PM) */
+    time: PropTypes.object,
     /** Callback function when the change event fires on the component */
     onChange: PropTypes.func
 };